moved grid_pixel_centres_2d_from
