+++ test_autoarray/geometry/test_geometry_util.py
@@ -1132,3 +1132,56 @@ def test__grid_pixels_2d_slim_from():
     )
 
     assert (grid_pixels == np.array([[1, 1], [1, 2], [2, 1], [2, 2]])).all()
+
+
+def test__grid_pixel_centres_2d_slim_from():
+
+    # coordinates in centres of pixels
+
+    grid_scaled = np.array([[1.0, -2.0], [1.0, 2.0], [-1.0, -2.0], [-1.0, 2.0]])
+
+    grid_pixels = aa.util.geometry.grid_pixel_centres_2d_slim_from(
+        grid_scaled_2d_slim=grid_scaled,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+    )
+
+    assert (grid_pixels == np.array([[0, 0], [0, 1], [1, 0], [1, 1]])).all()
+
+    # coordinates top-left of pixels
+
+    grid_scaled = np.array([[1.99, -3.99], [1.99, 0.01], [-0.01, -3.99], [-0.01, 0.01]])
+
+    grid_pixels = aa.util.geometry.grid_pixel_centres_2d_slim_from(
+        grid_scaled_2d_slim=grid_scaled,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+    )
+
+    assert (grid_pixels == np.array([[0, 0], [0, 1], [1, 0], [1, 1]])).all()
+
+    # coordinates bottom-right of pixels
+
+    grid_scaled = np.array([[0.01, -0.01], [0.01, 3.99], [-1.99, -0.01], [-1.99, 3.99]])
+
+    grid_pixels = aa.util.geometry.grid_pixel_centres_2d_slim_from(
+        grid_scaled_2d_slim=grid_scaled,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+    )
+
+    assert (grid_pixels == np.array([[0, 0], [0, 1], [1, 0], [1, 1]])).all()
+
+    # nonzero origin
+
+    # +1.0 for all entries for a origin of (1.0, 1.0)
+    grid_scaled = np.array([[2.0, -1.0], [2.0, 3.0], [0.0, -1.0], [0.0, 3.0]])
+
+    grid_pixels = aa.util.geometry.grid_pixel_centres_2d_slim_from(
+        grid_scaled_2d_slim=grid_scaled,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+        origin=(1.0, 1.0),
+    )
+
+    assert (grid_pixels == np.array([[0, 0], [0, 1], [1, 0], [1, 1]])).all()
